pyproject.toml
[tool.poetry]
name = "aws-auto-tagger"
version = "1.0.0"
description = "AWS Lambda service for automatically tagging AWS resources such as EC2, ELB, ENIs etc."
authors = ["Neil Camara"]

[tool.poetry.dependencies]
python = "^3.7"
boto3 = "^1.10"
"aws-cdk.aws-events" = "^1.19"
"aws-cdk.aws-events-targets" = "^1.19"
"aws-cdk.aws-lambda" = "^1.19"
"aws-cdk.core" = "^1.19"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
